AAPT/aapt_framework/orchestrator/requirements.txt
pika==1.3.1
Flask==2.3.2
neo4j==5.14.1
llama-cpp-python==0.2.23
requests==2.31.0 